Python/Generative_AI/Phidata-demo/requirements.txt
duckduckgo-search>=7.1.1
fastapi[standard]>=0.115.6
groq>=0.13.1
packaging>=24.2
phidata>=2.7.5
sqlalchemy>=2.0.36
uvicorn>=0.34.0
yfinance>=0.2.51
streamlit